src/agentic/requirements.txt
crewai
crewai-tools
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
qdrant-client
openai
